requirements.txt
numpy
matplotlib
scipy
python-dotenv
openai
webvtt-py
